Create bubble sort algorithm
I opted for "each" method since this was an enumerable that should not
change values, only indexes of those values. For efficiency I made sure
to both only look to the last sort - 1, and to lower the size by 1 at
the end. This gave it, at worst, a 50% increase in sorting speed.